# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VERGA API",
        "description": "API REST pour les applications externes VERGA (back-office agence Angular, application client mobile/web). Authentification Bearer Sanctum.",
        "contact": {
            "name": "VERGA",
            "email": "contact@verga.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/client/offres": {
            "get": {
                "tags": [
                    "Client - Offres"
                ],
                "summary": "Lister les offres disponibles",
                "description": "Catalogue public des offres actives avec capacité disponible > 0. Aucune authentification requise.",
                "operationId": "clientListOffres",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche titre, origine, destination",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "particulier",
                                "metre_cube",
                                "conteneur"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des offres actives"
                    }
                }
            }
        },
        "/client/commandes": {
            "get": {
                "tags": [
                    "Client - Commandes"
                ],
                "summary": "Lister mes commandes",
                "operationId": "clientListCommandes",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par code",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en_attente",
                                "confirmée",
                                "annulée"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des commandes du client"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Client - Commandes"
                ],
                "summary": "Créer une commande et initier le paiement",
                "description": "Crée en une seule requête : commande, colis (description + photos), paiement en attente, puis initie Bamboo Pay (redirection). Auth optionnelle : si un token Bearer est fourni, la commande est liée au compte client ; sinon commande invité (nom/prénom obligatoires). Réponse = lien de paiement + URL de vérification du statut.",
                "operationId": "clientCreateCommande",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "offre_id",
                                    "quantite",
                                    "telephone"
                                ],
                                "properties": {
                                    "offre_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "quantite": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Vêtements et chaussures",
                                        "nullable": true
                                    },
                                    "photos": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        },
                                        "maxItems": 5
                                    },
                                    "nom": {
                                        "description": "Obligatoire si invité (sans token)",
                                        "type": "string",
                                        "example": "Obame"
                                    },
                                    "prenom": {
                                        "description": "Obligatoire si invité (sans token)",
                                        "type": "string",
                                        "example": "Sarah"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "0612345678"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "offre_id",
                                    "quantite",
                                    "telephone",
                                    "nom",
                                    "prenom"
                                ],
                                "properties": {
                                    "offre_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "quantite": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    },
                                    "telephone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commande créée + liens de paiement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CheckoutResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation échouée ou stock insuffisant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/client/paiements/{code}/statut": {
            "get": {
                "tags": [
                    "Client - Paiements"
                ],
                "summary": "Vérifier le statut d'un paiement",
                "description": "Le client envoie la référence paiement VERGA (`PAY-...`). Le backend interroge Bamboo Pay si nécessaire et applique le règlement idempotent (même logique que le webhook callback).",
                "operationId": "clientCheckPaymentStatus",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Code paiement VERGA",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "PAY-ABCDEFGH"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut actuel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentStatusCheckResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paiement introuvable"
                    }
                }
            }
        },
...
        "/client/login": {
            "post": {
                "tags": [
                    "Client - Auth"
                ],
                "summary": "Connexion client",
                "operationId": "clientLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "device_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Compte bloqué"
                    },
                    "422": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/client/logout": {
            "post": {
                "tags": [
                    "Client - Auth"
                ],
                "summary": "Déconnexion client",
                "operationId": "clientLogout",
                "responses": {
                    "200": {
                        "description": "Token révoqué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/client/me": {
            "get": {
                "tags": [
                    "Client - Auth"
                ],
                "summary": "Profil client connecté",
                "operationId": "clientMe",
                "responses": {
                    "200": {
                        "description": "Profil utilisateur et client"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/client/commandes/{commande}": {
            "get": {
                "tags": [
                    "Client - Commandes"
                ],
                "summary": "Détail d'une commande",
                "operationId": "clientShowCommande",
                "parameters": [
                    {
                        "name": "commande",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail avec agence, offre, paiement, colis"
                    },
                    "404": {
                        "description": "Commande introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TokenResponse": {
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "1|abcdef..."
                    },
                    "token_type": {
                        "type": "string",
                        "example": "Bearer"
                    }
                },
                "type": "object"
            },
...
            "CheckoutResponse": {
                "properties": {
                    "commande_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "code": {
                        "type": "string",
                        "example": "CMD-ABCDEFGH"
                    },
                    "montant_total": {
                        "type": "number",
                        "format": "float",
                        "example": 25000
                    },
                    "paiement_code": {
                        "type": "string",
                        "example": "PAY-ABCDEFGH"
                    },
                    "redirect_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://devfront-bamboopay.ventis.group/pay/abc"
                    },
                    "verification_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "http://localhost/api/v1/client/paiements/PAY-ABCDEFGH/statut"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Sanctum obtenu via login ou register",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}